+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Upcoming
 * Add parameters to #bbcode_check_validity to add/remove tags
 * Support regular tag parameters (`[tag param=value][/tag]`) instead of 'quick parameters' (`[tag=value][/tag]`)
 * Changed tag description symbols (to become descriptive), **breaks existing custom tag additiona!**
+* Add support to show the BBCode annotated with errors (when there are any)
 
 Version 1.0.1 - 04-Jan-2015